# projects/data-quality-contracts-system/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Data Quality & Contracts System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/scenarios/{scenario_id}": {
      "get": {
        "summary": "Scenario",
        "operationId": "scenario_api_scenarios__scenario_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{scenario_id}/validate": {
      "post": {
        "summary": "Scenario Validate",
        "operationId": "scenario_validate_api_scenarios__scenario_id__validate_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationBundle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{scenario_id}/preview": {
      "post": {
        "summary": "Preview",
        "operationId": "preview_api_scenarios__scenario_id__preview_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/validate": {
      "post": {
        "summary": "External Validate",
        "operationId": "external_validate_api_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationBundle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compatibility": {
      "post": {
        "summary": "Compatibility",
        "operationId": "compatibility_api_compatibility_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompatibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/adapters/modeling": {
      "post": {
        "summary": "Modeling",
        "operationId": "modeling_api_adapters_modeling_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/adapters/observability": {
      "post": {
        "summary": "Observability",
        "operationId": "observability_api_adapters_observability_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CompatibilityRequest": {
        "properties": {
          "before": {
            "$ref": "#/components/schemas/DataContract-Input"
          },
          "after": {
            "$ref": "#/components/schemas/DataContract-Input"
          },
          "policy": {
            "$ref": "#/components/schemas/CompatibilityPolicy"
          },
          "compatible_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Dataset"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "before",
          "after"
        ],
        "title": "CompatibilityRequest"
      },
...
      "DataContract-Output": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "data-contract-v1",
            "title": "Contract Version",
            "default": "data-contract-v1"
          },
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Id"
          },
          "version": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Version",
            "default": "1.0"
          },
          "dataset_id": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Dataset Id"
          },
          "grain": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Grain"
          },
          "primary_key": {
            "items": {
              "type": "string",
              "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Primary Key"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/Column"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Columns"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/Relationship-Output"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Relationships"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/QualityRule-Output"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Rules"
          },
          "owner": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Owner"
          },
          "semantic_meaning": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Semantic Meaning"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "dataset_id",
          "grain",
          "primary_key",
          "columns",
          "owner",
          "semantic_meaning"
        ],
        "title": "DataContract"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QualityEvent": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "quality-event-v1",
            "title": "Contract Version",
            "default": "quality-event-v1"
          },
          "dataset_id": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Dataset Id"
          },
          "rule_id": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Rule Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "PASS",
              "WARN",
              "FAIL",
              "UNKNOWN"
            ],
            "title": "Status"
          },
          "severity": {
            "type": "string",
            "enum": [
              "INFO",
              "WARNING",
              "CRITICAL"
            ],
            "title": "Severity"
          },
          "expected": {
            "title": "Expected"
          },
          "actual": {
            "title": "Actual"
          },
          "failure_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Count"
          },
          "total_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Rows"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "input_fingerprint": {
            "type": "string",
            "title": "Input Fingerprint"
          },
          "evidence": {
            "$ref": "#/components/schemas/Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_id",
          "rule_id",
          "status",
          "severity",
          "expected",
          "actual",
          "failure_count",
          "total_rows",
          "timestamp",
          "run_id",
          "input_fingerprint",
          "evidence"
        ],
        "title": "QualityEvent"
      },
      "QualityGate": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "quality-gate-v1",
            "title": "Contract Version",
            "default": "quality-gate-v1"
          },
          "dataset_id": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Dataset Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "BLOCKED"
            ],
            "title": "Status"
          },
          "publication": {
            "type": "string",
            "enum": [
              "ELIGIBLE",
              "WITHHELD"
            ],
            "title": "Publication"
          },
          "blocking_rule_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocking Rule Ids"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "input_fingerprint": {
            "type": "string",
            "title": "Input Fingerprint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_id",
          "status",
          "publication",
          "blocking_rule_ids",
          "reasons",
          "run_id",
          "input_fingerprint"
        ],
        "title": "QualityGate"
      },
      "QualityRule-Input": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Description"
          },
          "target": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Target"
          },
          "target_type": {
            "type": "string",
            "enum": [
              "dataset",
              "column"
            ],
            "title": "Target Type",
            "default": "dataset"
          },
          "dimension": {
            "type": "string",
            "enum": [
              "schema",
              "completeness",
              "uniqueness",
              "validity",
              "consistency",
              "referential_integrity",
              "freshness",
              "volume",
              "business_rule",
              "reconciliation"
            ],
            "title": "Dimension"
          },
          "expectation": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Unique"
              },
              {
                "$ref": "#/components/schemas/NotNull-Input"
              },
              {
                "$ref": "#/components/schemas/AcceptedValues"
              },
              {
                "$ref": "#/components/schemas/Relationship-Input"
              },
              {
                "$ref": "#/components/schemas/Schema"
              },
              {
                "$ref": "#/components/schemas/Business"
              },
              {
                "$ref": "#/components/schemas/Freshness"
              },
              {
                "$ref": "#/components/schemas/Volume"
              },
              {
                "$ref": "#/components/schemas/Reconciliation-Input"
              }
            ],
            "title": "Expectation",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "accepted_values": "#/components/schemas/AcceptedValues",
                "business_rule": "#/components/schemas/Business",
                "freshness": "#/components/schemas/Freshness",
                "not_null": "#/components/schemas/NotNull-Input",
                "reconciliation": "#/components/schemas/Reconciliation-Input",
                "relationship": "#/components/schemas/Relationship-Input",
                "schema": "#/components/schemas/Schema",
                "unique": "#/components/schemas/Unique",
                "volume": "#/components/schemas/Volume"
              }
            }
          },
          "severity": {
            "type": "string",
            "enum": [
              "INFO",
              "WARNING",
              "CRITICAL"
            ],
            "title": "Severity",
            "default": "CRITICAL"
          },
          "blocking": {
            "type": "boolean",
            "title": "Blocking",
            "default": true
          },
          "owner": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Owner",
            "default": "Commerce Analytics"
          },
          "evidence_source": {
            "type": "string",
            "title": "Evidence Source",
            "default": "deterministic_dataset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "target",
          "dimension",
          "expectation"
        ],
        "title": "QualityRule"
      },
      "QualityRule-Output": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Description"
          },
          "target": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Target"
          },
          "target_type": {
            "type": "string",
            "enum": [
              "dataset",
              "column"
            ],
            "title": "Target Type",
            "default": "dataset"
          },
          "dimension": {
            "type": "string",
            "enum": [
              "schema",
              "completeness",
              "uniqueness",
              "validity",
              "consistency",
              "referential_integrity",
              "freshness",
              "volume",
              "business_rule",
              "reconciliation"
            ],
            "title": "Dimension"
          },
          "expectation": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Unique"
              },
              {
                "$ref": "#/components/schemas/NotNull-Output"
              },
              {
                "$ref": "#/components/schemas/AcceptedValues"
              },
              {
                "$ref": "#/components/schemas/Relationship-Output"
              },
              {
                "$ref": "#/components/schemas/Schema"
              },
              {
                "$ref": "#/components/schemas/Business"
              },
              {
                "$ref": "#/components/schemas/Freshness"
              },
              {
                "$ref": "#/components/schemas/Volume"
              },
              {
                "$ref": "#/components/schemas/Reconciliation-Output"
              }
            ],
            "title": "Expectation",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "accepted_values": "#/components/schemas/AcceptedValues",
                "business_rule": "#/components/schemas/Business",
                "freshness": "#/components/schemas/Freshness",
                "not_null": "#/components/schemas/NotNull-Output",
                "reconciliation": "#/components/schemas/Reconciliation-Output",
                "relationship": "#/components/schemas/Relationship-Output",
                "schema": "#/components/schemas/Schema",
                "unique": "#/components/schemas/Unique",
                "volume": "#/components/schemas/Volume"
              }
            }
          },
          "severity": {
            "type": "string",
            "enum": [
              "INFO",
              "WARNING",
              "CRITICAL"
            ],
            "title": "Severity",
            "default": "CRITICAL"
          },
          "blocking": {
            "type": "boolean",
            "title": "Blocking",
            "default": true
          },
          "owner": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Owner",
            "default": "Commerce Analytics"
          },
          "evidence_source": {
            "type": "string",
            "title": "Evidence Source",
            "default": "deterministic_dataset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "target",
          "dimension",
          "expectation"
        ],
        "title": "QualityRule"
      },
...
      "ScenarioRequest": {
        "properties": {
          "corruptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Corruptions"
          },
          "executed_at": {
            "type": "string",
            "title": "Executed At",
            "default": "2026-10-02T00:00:00Z"
          },
          "run_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Run Id",
            "default": "commerce-run-1"
          },
          "additional_rules": {
            "items": {
              "$ref": "#/components/schemas/QualityRule-Input"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Additional Rules"
          },
          "rule_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ScenarioRequest"
      },
      "Schema": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "schema",
            "title": "Kind",
            "default": "schema"
          },
          "allow_extra_columns": {
            "type": "boolean",
            "title": "Allow Extra Columns",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Schema"
      },
...
      "ValidateRequest": {
        "properties": {
          "contract": {
            "$ref": "#/components/schemas/DataContract-Input"
          },
          "datasets": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Dataset"
            },
            "type": "object",
            "maxProperties": 20,
            "title": "Datasets"
          },
          "executed_at": {
            "type": "string",
            "title": "Executed At"
          },
          "run_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Run Id"
          },
          "additional_rules": {
            "items": {
              "$ref": "#/components/schemas/QualityRule-Input"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Additional Rules"
          },
          "rule_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$"
                },
                "type": "array",
                "maxItems": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contract",
          "datasets",
          "executed_at",
          "run_id"
        ],
        "title": "ValidateRequest"
      },
      "ValidationBundle": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "quality-bundle-v1",
            "title": "Contract Version",
            "default": "quality-bundle-v1"
          },
          "contract": {
            "$ref": "#/components/schemas/DataContract-Output"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/QualityRule-Output"
            },
            "type": "array",
            "title": "Rules"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ValidationResult"
            },
            "type": "array",
            "title": "Results"
          },
          "gate": {
            "$ref": "#/components/schemas/QualityGate"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/QualityEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Summary"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "input_fingerprint": {
            "type": "string",
            "title": "Input Fingerprint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contract",
          "rules",
          "results",
          "gate",
          "events",
          "summary",
          "run_id",
          "input_fingerprint"
        ],
        "title": "ValidationBundle"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationResult": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "validation-result-v1",
            "title": "Contract Version",
            "default": "validation-result-v1"
          },
          "rule_id": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Rule Id"
          },
          "target_id": {
            "type": "string",
            "pattern": "^[A-Za-z_][A-Za-z0-9_]{0,79}$",
            "title": "Target Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "PASS",
              "WARN",
              "FAIL",
              "UNKNOWN"
            ],
            "title": "Status"
          },
          "expected": {
            "title": "Expected"
          },
          "actual": {
            "title": "Actual"
          },
          "failed_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed Rows"
          },
          "total_rows": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Rows"
          },
          "failure_rate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Rate"
          },
          "severity": {
            "type": "string",
            "enum": [
              "INFO",
              "WARNING",
              "CRITICAL"
            ],
            "title": "Severity"
          },
          "blocking": {
            "type": "boolean",
            "title": "Blocking"
          },
          "executed_at": {
            "type": "string",
            "title": "Executed At"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "input_fingerprint": {
            "type": "string",
            "title": "Input Fingerprint"
          },
          "evidence": {
            "$ref": "#/components/schemas/Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rule_id",
          "target_id",
          "status",
          "expected",
          "actual",
          "failed_rows",
          "total_rows",
          "failure_rate",
          "severity",
          "blocking",
          "executed_at",
          "run_id",
          "input_fingerprint",
          "evidence"
        ],
        "title": "ValidationResult"
      },
...
    }
  }
}